sparc: fix #1061

+++ arch/Sparc/SparcInstPrinter.c
@@ -269,7 +269,7 @@ static void printOperand(MCInst *MI, int opNum, SStream *O)
 				SStream_concat(O, "%u", Imm);
 		} else {
 			if (Imm < -HEX_THRESHOLD)
-				SStream_concat(O, "-0x%x", -Imm);
+				SStream_concat(O, "-0x%x", (unsigned int)-Imm);
 			else
 				SStream_concat(O, "-%u", -Imm);
 		}
